+++ reset_select_language_entries.rb
@@ -3,7 +3,7 @@
 require "json"
 
 language_files = Dir["ValheimArmory/localizations/*"]
-keys_to_remove = %w[item_meteor_atgeir]
+keys_to_remove = %w[item_battleaxe_blackmetal item_battleaxe_blackmetal_description]
 
 language_files.each do |lang_file|
   next if lang_file == "ValheimArmory/localizations/English.json"